Fixed a logging issue and tweaked regex
There was an issue within the main function that when an error occured during the URLCheck, it would fatal out and exit. This has been removed due to double error checking inside and outside of the function.

I have also included the outter bracket of the regex to escape the final characters of the markdown URL string.